fix process! startup on Julia 1.0, 32 bit
see [issue #8](https://github.com/pbayer/Simulate.jl/issues/8)

+++ src/process.jl
@@ -39,7 +39,7 @@ end
 Start a `SimProcess` as a task in a loop.
 """
 function startup!(p::SimProcess, cycles::Number)
-    start = Channel{Int64}(0)
+    start = Channel{Int}(0)
     p.task = @async loop(p, start, cycles)
     p.state = Idle()
     put!(start, 1) # let the process start
